+++ src/cpu.c
@@ -68,7 +68,7 @@ const unsigned char flagCarry = (1 << 4);
 static char* lastOpcodeName = "";
 static unsigned short lastOperand = 0;
 
-unsigned int tickCounter = 0;
+unsigned long int tickCounter = 0;
 unsigned char stopped = 0;
 
 void setFlag(unsigned char flag) { registers.F |= flag; }
@@ -220,45 +220,9 @@ int cpuStep(void) {
   printf("\n*********************************\n");
   // #endif
   int ticksUsed = tickCounter - tickStart;
-  // printf("Ticks used = %d\n", ticksUsed);
-  // char c = getchar();
 
-  // if (cycles >= 12375) {
-  //   // if (1) {
-  //   printf("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n");
-  //   printf(
-  //       "Registers\nA: 0x%02X B: 0x%02X C: 0x%02X D: 0x%02X\nE: 0x%02X F: "
-  //       "0x%02X "
-  //       "H: 0x%02X L: 0x%02X\n",
-  //       registers.A, registers.B, registers.C, registers.D, registers.E,
-  //       registers.F, registers.H, registers.L);
-  //   printf("Flags: Z: %d N: %d H: %d C: %d\n", checkFlag(flagZero),
-  //          checkFlag(flagNegative), checkFlag(flagHalfCarry),
-  //          checkFlag(flagCarry));
-  //   printf("PC: 0x%02X SP: 0x%02X\n", registers.PC, registers.SP);
-  //   printf("Total cycles (ticks): %d\n", tickCounter);
-  //   printf("\n");
-  //   char c = getchar();
-  // }
 
   ++cycles;
-
-  // opcode aNextOpcode;
-  // unsigned char aNextInstruction = mmuReadByte(registers.PC);
-  // if (aNextInstruction == 0xCB) {
-  //   aNextInstruction = mmuReadByte(registers.PC + 1);
-  //   aNextOpcode = CBOpcodeTable[aNextInstruction];
-  // } else {
-  //   aNextOpcode = baseOpcodeTable[aNextInstruction];
-  // }
-  // if (!strcmp(aOpcode.asmName, "RST 38") && !strcmp(aNextOpcode.asmName, "RST 38")) {
-  //   printf("RST 38 loop caught\n");
-  //   while (1) {
-  //     char aCrap[10];
-  //     fgets(aCrap, 9, stdin);
-  //   }
-  // }
-
 
   return ticksUsed;
 }
